Whole: compile_commands and compile errors
Created a compile commands, includes in the bootstrapper can now look
wonderful and not be a mess of local paths. I am pretty sure that there
is some type of error when compiling the kernel. For some reason
defining an array of 16 gdt_entries in the kernel causes the entire
thing to fall over when we call an interrupt. Whatever is causing this
can most likely also be linked to what is causing the processor to
double fault with an int3.

+++ bootstrap/src/c/include/elf/elf.h
@@ -1,8 +1,8 @@
 #ifndef ARC_ELF_ELF_H
 #define ARC_ELF_ELF_H
 
-#include "../global.h"
+#include <global.h>
 
-int load_elf(uint64_t *pml4, void *elf);
+uint64_t load_elf(uint64_t *pml4, void *elf);
 
 #endif
